ref: cast less often

+++ plight/include/component/render_data.h
@@ -14,7 +14,7 @@ namespace Plight::Component
         unsigned int            m_shaderProgramId = 0;
 
         // Total number of visits to the drawn vertices (some vertices are visited multiple times while drawing)
-        size_t                  m_vertexVisits = 0;
+        int                     m_vertexVisits = 0;
 
         // OpenGL-internal vertex array object id
         unsigned int            m_vertexArrayObject = 0;
